// src/donation/donation.types.ts
export enum DonationStatus {
  PENDING = 'pending',
  COMPLETED = 'completed',
}

export enum EvaluationSatisfaction {
  NEUTRAL = 'neutral',
  HAPPY = 'happy',
  THRILLED = 'thrilled',
}

export enum EvaluationBonus {
  NONE = 'none',
  ON_TIME = 'on_time',
  WENT_ABOVE_AND_BEYOND = 'went_above_and_beyond',
}